requirements.txt
Flask==3.0.0
google-cloud-bigquery==3.13.0
tabulate==0.9.0
requests==2.13.0
gunicorn==21.2.0